requirements.txt
PyJWT
schema